Found a float conversion bug, fixed it

+++ robogen/src/evolution/representation/RobotRepresentation.cpp
@@ -1391,7 +1391,7 @@ float RobotRepresentation::calculateBrainComplexity(){
 void RobotRepresentation::calculateRobotComplexity(){
 	float body = (float) (this->calculateBodyComplexity(bodyTree_) - MIN_BODY_COMPLEXITY)/(MAX_BODY_COMPLEXITY - MIN_BODY_COMPLEXITY);
 	float brain = this->calculateBrainComplexity();
-	complexity_ = (double) (0.7*body + 0.3*brain);
+	complexity_ = (float) (0.7*body + 0.3*brain);
 }
 
 // CH - returns complexity of robot
